+++ tests/ManagerTest.php
@@ -9,9 +9,8 @@ use Innmind\Doctrine\{
     Exception\NestedMutationNotSupported,
     Exception\MutationOutsideOfContext,
 };
-use Doctrine\ORM\{
-    EntityManagerInterface,
-};
+use Innmind\Immutable\Either;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ObjectRepository;
 use PHPUnit\Framework\TestCase;
 use Innmind\BlackBox\{
@@ -63,9 +62,18 @@ class ManagerTest extends TestCase
     {
         $manager = Manager::of($this->createMock(EntityManagerInterface::class));
 
-        $this->assertNull($manager->mutate(function($inner) use ($manager) {
-            $this->assertSame($manager, $inner);
-        }));
+        $this->assertTrue(
+            $manager
+                ->mutate(function($inner) use ($manager) {
+                    $this->assertSame($manager, $inner);
+
+                    return Either::right(null);
+                })
+                ->match(
+                    static fn() => true,
+                    static fn() => false,
+                ),
+        );
     }
 
     public function testFlushOnceTheMutationIsDone()
@@ -77,9 +85,21 @@ class ManagerTest extends TestCase
             ->expects($this->exactly(2))
             ->method('flush');
 
-        $this->assertNull($manager->mutate(static fn() => null));
-        $this->assertNull(
-            $manager->mutate(static fn() => null),
+        $this->assertTrue(
+            $manager
+                ->mutate(static fn() => Either::right(null))
+                ->match(
+                    static fn() => true,
+                    static fn() => false,
+                ),
+        );
+        $this->assertTrue(
+            $manager
+                ->mutate(static fn() => Either::right(null))
+                ->match(
+                    static fn() => true,
+                    static fn() => false,
+                ),
             'Multiple mutations should be allowed',
         );
     }
@@ -95,7 +115,39 @@ class ManagerTest extends TestCase
 
                 $this->assertSame(
                     $return,
-                    $manager->mutate(static fn() => $return),
+                    $manager
+                        ->mutate(static fn() => Either::right($return))
+                        ->match(
+                            static fn($value) => $value,
+                            static fn() => null,
+                        ),
+                );
+            });
+    }
+
+    public function testCloseTheEntityManagerWhenMutationReturnsAnError()
+    {
+        $this
+            ->forAll(Set\AnyType::any())
+            ->then(function($return) {
+                $manager = Manager::of(
+                    $em = $this->createMock(EntityManagerInterface::class),
+                );
+                $em
+                    ->expects($this->once())
+                    ->method('close');
+                $return = new \Exception;
+
+                $this->assertSame(
+                    $return,
+                    $manager
+                        ->mutate(static function() use ($return) {
+                            return Either::left($return);
+                        })
+                        ->match(
+                            static fn() => null,
+                            static fn($e) => $e,
+                        ),
                 );
             });
     }
@@ -151,6 +203,8 @@ class ManagerTest extends TestCase
                     $repository = $manager->repository($entityClass);
 
                     $this->assertNull($repository->add($entity));
+
+                    return Either::right(null);
                 });
             });
     }
@@ -186,8 +240,13 @@ class ManagerTest extends TestCase
             $this->fail('it should throw');
         } catch (\Throwable $e) {
             $this->assertSame($exception, $e);
-            $this->assertNull(
-                $manager->transaction(static fn() => null),
+            $this->assertTrue(
+                $manager
+                    ->transaction(static fn() => Either::right(null))
+                    ->match(
+                        static fn() => true,
+                        static fn() => false,
+                    ),
                 'the manager should be healthy after a failed transaction',
             );
         }
@@ -213,10 +272,20 @@ class ManagerTest extends TestCase
 
                 $this->assertSame(
                     $return,
-                    $manager->transaction(static fn() => $return),
+                    $manager
+                        ->transaction(static fn() => Either::right($return))
+                        ->match(
+                            static fn($value) => $value,
+                            static fn() => null,
+                        ),
                 );
-                $this->assertNull(
-                    $manager->transaction(static fn() => null),
+                $this->assertTrue(
+                    $manager
+                        ->transaction(static fn() => Either::right(null))
+                        ->match(
+                            static fn() => true,
+                            static fn() => false,
+                        ),
                     'the manager should be healthy after a transaction',
                 );
             });
@@ -226,9 +295,18 @@ class ManagerTest extends TestCase
     {
         $manager = Manager::of($this->createMock(EntityManagerInterface::class));
 
-        $this->assertNull($manager->transaction(function($inner) use ($manager) {
-            $this->assertSame($manager, $inner);
-        }));
+        $this->assertTrue(
+            $manager
+                ->transaction(function($inner) use ($manager) {
+                    $this->assertSame($manager, $inner);
+
+                    return Either::right(null);
+                })
+                ->match(
+                    static fn() => true,
+                    static fn() => false,
+                ),
+        );
     }
 
     public function testAllowMutationInTransaction()
@@ -245,6 +323,8 @@ class ManagerTest extends TestCase
                     $repository = $manager->repository($entityClass);
 
                     $this->assertNull($repository->add($entity));
+
+                    return Either::right(null);
                 });
             });
     }
@@ -267,6 +347,6 @@ class ManagerTest extends TestCase
             ->expects($this->once())
             ->method('commit');
 
-        $manager->transaction(static fn($_, $flush) => $flush());
+        $manager->transaction(static fn($_, $flush) => Either::right($flush()));
     }
 }